use num proposals to measure deadline

+++ packages/protocol/script/layer1/based/DeployProtocolOnL1.s.sol
@@ -413,7 +413,7 @@ contract DeployProtocolOnL1 is DeployCapability {
             impl: address(
                 new ForcedInclusionStore(
                     resolver,
-                    vm.envUint("INCLUSION_WINDOW"),
+                    uint8(vm.envUint("INCLUSION_WINDOW")),
                     uint64(vm.envUint("INCLUSION_FEE_IN_GWEI"))
                 )
             ),
